Separate AudioData from MicrophoneManager
